refactor(bridge): make useTransactionHistory the single tx history entry point
useTransactionHistory now takes an object param with an optional txHash.
When txHash is set, the address pipeline is disabled and the hash is
resolved directly. Removes the useDisplayedTransactionHistory wrapper.
