+++ packages/arb-token-bridge-ui/src/components/Widget/WidgetTransactionHistory.tsx
@@ -28,7 +28,9 @@ const WidgetTransactionHistoryLoadingPlaceholder = () => {
 
 export const WidgetTransactionHistory = (props: UseDialogProps) => {
   const { address: walletAddress } = useAccount();
-  const { transactions, loading: isLoadingTransactions } = useTransactionHistory(walletAddress);
+  const { transactions, loading: isLoadingTransactions } = useTransactionHistory({
+    address: walletAddress,
+  });
   const currentTx = transactions[0];
   const isLoading = transactions.length ? false : isLoadingTransactions;
